# src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "消防巡检管理系统 API",
    "description": "基于 Node.js + Express + MongoDB 的消防设备巡检、报警处理、安全管理系统 API 文档。\n\n## 认证方式\n所有私有接口需在请求头中携带 Authorization: Bearer <token>。\n\n## 数据范围\n部分接口受数据范围权限控制：all（全部）/ department（本部门）/ self（仅本人）/ none（无权限）。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "开发环境"
    },
    {
      "url": "/",
      "description": "当前服务器"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/roles": {
      "get": {
        "tags": ["Roles"],
        "summary": "获取角色列表（分页）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": ["Roles"],
        "summary": "创建角色",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "code"],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^[A-Z_]+$"
                  },
                  "description": {
                    "type": "string"
                  },
                  "level": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 10
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功"
          }
        }
      }
    },
...
    "/api/security/ip-list": {
      "get": {
        "tags": ["Security"],
        "summary": "获取 IP 黑白名单列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["black", "white"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "post": {
        "tags": ["Security"],
        "summary": "添加 IP 到黑/白名单",
        "description": "白名单优先级高于黑名单；加入黑名单时若已在白名单则拒绝",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ip"],
                "properties": {
                  "ip": {
                    "type": "string",
                    "description": "支持 IPv4/IPv6/CIDR"
                  },
                  "type": {
                    "type": "string",
                    "enum": ["black", "white"],
                    "default": "black"
                  },
                  "reason": {
                    "type": "string",
                    "maxLength": 200
                  },
                  "durationHours": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 8760,
                    "description": "0=永久"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "添加成功"
          },
          "400": {
            "description": "IP 已在白名单中"
          }
        }
      }
    },
    "/api/security/ip-list/{id}": {
      "delete": {
        "tags": ["Security"],
        "summary": "从名单中移除 IP 记录",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
  }
}